Reintroduce GSym constructor instead of using Core.Var
Because we don't actually assign anything to the placeholder
variables, they don't really behave as the normal Core.Var
types. We don't define these when constructing a node,
and so nothing is really bound to them.

This messes up the clock checker, so we reintroduce the use
of the custom GSym constructor and clock check them in a
hand-wavy "don't worry about it" manner. Might work, we'll
see.