Add helper function for FinEnum identity instances
Since a `Partition` instance is necessary to use `caseof`,
it was a bit inconvenient for non-"normal" sum types,
e.g. Int8. The helper function `enumIdPartitions` acts similar
to The Trick, enumerating through all values of the input
type of the `Partition` to generate literal constructor cases.

By implementing `partition` with this for instances such as
`Partition Int8 Int8`, we can get similar functionality to
that of `match`.